refactor: add LotCurrentBidInfo component
